+++ src/main/java/tornadofx/App.kt
@@ -1,10 +1,20 @@
 package tornadofx
 
 import javafx.application.Application
+import javafx.application.Platform
 import javafx.scene.Scene
 import javafx.scene.image.Image
 import javafx.scene.layout.Pane
 import javafx.stage.Stage
+import java.awt.*
+import java.awt.event.MouseAdapter
+import java.awt.event.MouseEvent
+import java.awt.event.MouseEvent.BUTTON1
+import java.awt.image.BufferedImage
+import java.io.InputStream
+import java.util.*
+import javax.imageio.ImageIO
+import javax.swing.SwingUtilities
 import kotlin.properties.ReadOnlyProperty
 import kotlin.reflect.KClass
 import kotlin.reflect.KProperty
@@ -12,6 +22,10 @@ import kotlin.reflect.KProperty
 open class App(primaryView: KClass<out UIComponent>? = null, vararg stylesheet: KClass<out Stylesheet>) : Application() {
     var scope: Scope = DefaultScope
     val workspace: Workspace get() = scope.workspace
+    private val trayIcons = ArrayList<TrayIcon>()
+    val resources: ResourceLookup by lazy {
+        ResourceLookup(this)
+    }
 
     fun <T : FXEvent> fire(event: T) {
         FX.eventbus.fire(event)
@@ -65,6 +79,9 @@ open class App(primaryView: KClass<out UIComponent>? = null, vararg stylesheet: 
 
     override fun stop() {
         scope.deregister()
+        trayIcons.forEach {
+            SwingUtilities.invokeLater { SystemTray.getSystemTray().remove(it) }
+        }
     }
 
     open fun shouldShowPrimaryStage() = true
@@ -90,6 +107,61 @@ open class App(primaryView: KClass<out UIComponent>? = null, vararg stylesheet: 
 
     class DeterminedByParameter : View() {
         override val root = Pane()
+    }
+
+    fun trayicon(image: BufferedImage, tooltip: String?, implicitExit: Boolean = false, autoSize: Boolean = false, op: TrayIcon.() -> Unit){
+        Platform.setImplicitExit(implicitExit)
+        SwingUtilities.invokeLater {
+            Toolkit.getDefaultToolkit()
+            val trayIcon = TrayIcon(image, tooltip)
+            trayIcon.isImageAutoSize= autoSize
+            op(trayIcon)
+            SystemTray.getSystemTray().add(trayIcon)
+            trayIcons.add(trayIcon)
+        }
+    }
+    fun trayicon(icon: InputStream, tooltip: String? = null, implicitExit: Boolean = false, autoSize: Boolean = false, op: TrayIcon.() -> Unit) {
+        trayicon(ImageIO.read(icon), tooltip, implicitExit, autoSize, op)
+    }
+
+    fun TrayIcon.menu(label: String, op: PopupMenu.() -> Unit) {
+        popupMenu = PopupMenu(label)
+        op(popupMenu)
+    }
+
+    fun TrayIcon.setOnMouseClicked(fxThread: Boolean = false, button: Int = BUTTON1, clickCount: Int = 1, op: () -> Unit) {
+        addMouseListener(object : MouseAdapter() {
+            override fun mouseClicked(e: MouseEvent) {
+                if (e.button == button && e.clickCount == clickCount) {
+                    if (fxThread) {
+                        Platform.runLater {
+                            op()
+                        }
+                    } else {
+                        op()
+                    }
+                }
+            }
+        })
+    }
+
+    fun PopupMenu.item(label: String, shortcut: MenuShortcut? = null, op: MenuItem.() -> Unit): MenuItem {
+        val item = MenuItem(label, shortcut)
+        op(item)
+        add(item)
+        return item
+    }
+
+    fun MenuItem.setOnAction(fxThread: Boolean = false, action: () -> Unit) {
+        addActionListener {
+            if (fxThread) {
+                Platform.runLater {
+                    action()
+                }
+            } else {
+                action()
+            }
+        }
     }
 
 }